Inventory class

A list containing forest inventory data summaries. It includes data by diametric class (dclass_metrics) and by plot and species (group_metrics).

Usage

Inventory(dclass_metrics = data.frame(), group_metrics = data.frame())

Arguments

dclass_metrics

A data.frame summarised by diametric class with variables such as plot_id, species, dclass, height, ntrees, ntrees_ha, h0, dg, and g_ha.

group_metrics

A data.frame summarised by plot and species with variables such as plot_id, species, d_mean, d_median, d_sd, dg, h_mean, h_median, h_sd, h_lorey, h0, ntrees, ntrees_ha, g_ha, and spacing.

Value

An S7 Inventory object, which contains the inventory data.